Fix Play YouTube video button
Changing to an application means that Elm will always try to handle link
clicks, even if there's another click handler with prevent default. It
is possible to add a `target` to links that I want to handle myself, but
this really highlights that I shouldn't double purpose the link/button.

I have solved this by making "Play YouTube video" a button, with a
nearby link to view the video on YouTube.